+++ php-src/Rules/LesserThan.php
@@ -18,7 +18,7 @@ class LesserThan extends ARule
 
     public function validate(IValidate $entry): void
     {
-        if (intval($entry->getValue()) >= $this->againstValue) {
+        if (floatval($entry->getValue()) >= $this->againstValue) {
             throw new RuleException($this->errorText);
         }
     }
